changed baseTemplateService to standardTemplateService
