+++ libs/logic-apps-shared/src/designer-client-services/lib/base/index.ts
@@ -31,7 +31,6 @@ export * from './operations';
 export { BaseChatbotService } from './chatbot';
 export type { ChatbotServiceOptions } from './chatbot';
 // Template
-export { BaseTemplateService } from './template';
 export type { TemplateServiceOptions } from './template';
 // Tenant
 export { BaseTenantService } from './tenant';
