+++ libs/logic-apps-shared/src/designer-client-services/lib/standard/index.ts
@@ -5,3 +5,5 @@ export { StandardSearchService } from './search';
 export { StandardRunService } from './run';
 export { StandardArtifactService } from './artifact';
 export { StandardCustomCodeService } from './customcode';
+// Template
+export { StandardTemplateService } from './template';
